Refactoring and cleanup
- Updated error handling and the interface of Error.
- Updated Lazy<ProgVar>.
- Updated DSA texture and framebuffer.
- Updated several examples.
- Removed one of the older examples.
- Work-in-progress, some parts won't compile.

+++ implement/oglplus/error/object.ipp
@@ -73,9 +73,9 @@ const char* ObjectError::TargetName(void) const
 }
 
 OGLPLUS_LIB_FUNC
-GLuint ObjectError::ObjectName(void) const
+GLint ObjectError::ObjectName(void) const
 {
-	return _obj_name;
+	return GLint(_obj_name);
 }
 
 OGLPLUS_LIB_FUNC
@@ -119,9 +119,9 @@ const char* ObjectPairError::SubjectClassName(void) const
 }
 
 OGLPLUS_LIB_FUNC
-GLuint ObjectPairError::SubjectName(void) const
+GLint ObjectPairError::SubjectName(void) const
 {
-	return _sub_name;
+	return GLint(_sub_name);
 }
 
 OGLPLUS_LIB_FUNC
